search: test some error conditions of client.SearchByEntity
